impr

+++ Avalonia.NETCoreApp/Organista/Files/AudioFile.cs
@@ -7,7 +7,7 @@ namespace Organista
         public string path { get; set; }
         public string title { get; set; }
         public string album { get; set; }
-        public int length { get; set; }
+        public long length { get; set; }
         public TagValue tagValue { get; set; }
 
         public AudioFile()
