+++ Avalonia.NETCoreApp/Organista/Files/VideoFile.cs
@@ -5,7 +5,7 @@ namespace Organista
         public string path { get; set; }
         public string title { get; set; }
         public string folder { get; set; }
-        public int length { get; set; }
+        public long length { get; set; }
         public int height { get; set; }
         public int width { get; set; }
